# gen/demo/model/v1/tasks.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.12
// 	protoc        (unknown)
// source: demo/model/v1/tasks.proto

package modelv1

import (
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

// Task 是任务数据；时间使用 RFC 3339 JSON 字符串。
type Task struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Id            string                 `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
	Title         string                 `protobuf:"bytes,2,opt,name=title,proto3" json:"title,omitempty"`
	Description   string                 `protobuf:"bytes,3,opt,name=description,proto3" json:"description,omitempty"`
	Completed     bool                   `protobuf:"varint,4,opt,name=completed,proto3" json:"completed,omitempty"`
	CreatedAt     *timestamppb.Timestamp `protobuf:"bytes,5,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty"`
	UpdatedAt     *timestamppb.Timestamp `protobuf:"bytes,6,opt,name=updated_at,json=updatedAt,proto3" json:"updated_at,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *Task) Reset() {
	*x = Task{}
	mi := &file_demo_model_v1_tasks_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *Task) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*Task) ProtoMessage() {}

func (x *Task) ProtoReflect() protoreflect.Message {
	mi := &file_demo_model_v1_tasks_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use Task.ProtoReflect.Descriptor instead.
func (*Task) Descriptor() ([]byte, []int) {
	return file_demo_model_v1_tasks_proto_rawDescGZIP(), []int{0}
}

func (x *Task) GetId() string {
	if x != nil {
		return x.Id
	}
	return ""
}

func (x *Task) GetTitle() string {
	if x != nil {
		return x.Title
	}
	return ""
}

func (x *Task) GetDescription() string {
	if x != nil {
		return x.Description
	}
	return ""
}

func (x *Task) GetCompleted() bool {
	if x != nil {
		return x.Completed
	}
	return false
}

func (x *Task) GetCreatedAt() *timestamppb.Timestamp {
	if x != nil {
		return x.CreatedAt
	}
	return nil
}

func (x *Task) GetUpdatedAt() *timestamppb.Timestamp {
	if x != nil {
		return x.UpdatedAt
	}
	return nil
}

// TaskEvent 定义 SSE data 的负载，HTTP 的 SSE 分帧由手写 handler 负责。
type TaskEvent struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	EventId       string                 `protobuf:"bytes,1,opt,name=event_id,json=eventId,proto3" json:"event_id,omitempty"`
	Type          string                 `protobuf:"bytes,2,opt,name=type,proto3" json:"type,omitempty"`
	Task          *Task                  `protobuf:"bytes,3,opt,name=task,proto3" json:"task,omitempty"`
	OccurredAt    *timestamppb.Timestamp `protobuf:"bytes,4,opt,name=occurred_at,json=occurredAt,proto3" json:"occurred_at,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *TaskEvent) Reset() {
	*x = TaskEvent{}
	mi := &file_demo_model_v1_tasks_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *TaskEvent) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*TaskEvent) ProtoMessage() {}

func (x *TaskEvent) ProtoReflect() protoreflect.Message {
	mi := &file_demo_model_v1_tasks_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use TaskEvent.ProtoReflect.Descriptor instead.
func (*TaskEvent) Descriptor() ([]byte, []int) {
	return file_demo_model_v1_tasks_proto_rawDescGZIP(), []int{1}
}

func (x *TaskEvent) GetEventId() string {
	if x != nil {
		return x.EventId
	}
	return ""
}

func (x *TaskEvent) GetType() string {
	if x != nil {
		return x.Type
	}
	return ""
}

func (x *TaskEvent) GetTask() *Task {
	if x != nil {
		return x.Task
	}
	return nil
}

func (x *TaskEvent) GetOccurredAt() *timestamppb.Timestamp {
	if x != nil {
		return x.OccurredAt
	}
	return nil
}

var File_demo_model_v1_tasks_proto protoreflect.FileDescriptor

const file_demo_model_v1_tasks_proto_rawDesc = "" +
	"\n" +
	"\x19demo/model/v1/tasks.proto\x12\rdemo.model.v1\x1a\x1fgoogle/protobuf/timestamp.proto\"\xe2\x01\n" +
	"\x04Task\x12\x0e\n" +
	"\x02id\x18\x01 \x01(\tR\x02id\x12\x14\n" +
	"\x05title\x18\x02 \x01(\tR\x05title\x12 \n" +
	"\vdescription\x18\x03 \x01(\tR\vdescription\x12\x1c\n" +
	"\tcompleted\x18\x04 \x01(\bR\tcompleted\x129\n" +
	"\n" +
	"created_at\x18\x05 \x01(\v2\x1a.google.protobuf.TimestampR\tcreatedAt\x129\n" +
	"\n" +
	"updated_at\x18\x06 \x01(\v2\x1a.google.protobuf.TimestampR\tupdatedAt\"\xa0\x01\n" +
	"\tTaskEvent\x12\x19\n" +
	"\bevent_id\x18\x01 \x01(\tR\aeventId\x12\x12\n" +
	"\x04type\x18\x02 \x01(\tR\x04type\x12'\n" +
	"\x04task\x18\x03 \x01(\v2\x13.demo.model.v1.TaskR\x04task\x12;\n" +
	"\voccurred_at\x18\x04 \x01(\v2\x1a.google.protobuf.TimestampR\n" +
	"occurredAtB:Z8example.com/api-toolchain-demo/gen/demo/model/v1;modelv1b\x06proto3"

var (
	file_demo_model_v1_tasks_proto_rawDescOnce sync.Once
	file_demo_model_v1_tasks_proto_rawDescData []byte
)

func file_demo_model_v1_tasks_proto_rawDescGZIP() []byte {
	file_demo_model_v1_tasks_proto_rawDescOnce.Do(func() {
		file_demo_model_v1_tasks_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_demo_model_v1_tasks_proto_rawDesc), len(file_demo_model_v1_tasks_proto_rawDesc)))
	})
	return file_demo_model_v1_tasks_proto_rawDescData
}

var file_demo_model_v1_tasks_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_demo_model_v1_tasks_proto_goTypes = []any{
	(*Task)(nil),                  // 0: demo.model.v1.Task
	(*TaskEvent)(nil),             // 1: demo.model.v1.TaskEvent
	(*timestamppb.Timestamp)(nil), // 2: google.protobuf.Timestamp
}
var file_demo_model_v1_tasks_proto_depIdxs = []int32{
	2, // 0: demo.model.v1.Task.created_at:type_name -> google.protobuf.Timestamp
	2, // 1: demo.model.v1.Task.updated_at:type_name -> google.protobuf.Timestamp
	0, // 2: demo.model.v1.TaskEvent.task:type_name -> demo.model.v1.Task
	2, // 3: demo.model.v1.TaskEvent.occurred_at:type_name -> google.protobuf.Timestamp
	4, // [4:4] is the sub-list for method output_type
	4, // [4:4] is the sub-list for method input_type
	4, // [4:4] is the sub-list for extension type_name
	4, // [4:4] is the sub-list for extension extendee
	0, // [0:4] is the sub-list for field type_name
}

func init() { file_demo_model_v1_tasks_proto_init() }
func file_demo_model_v1_tasks_proto_init() {
	if File_demo_model_v1_tasks_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_demo_model_v1_tasks_proto_rawDesc), len(file_demo_model_v1_tasks_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   0,
		},
		GoTypes:           file_demo_model_v1_tasks_proto_goTypes,
		DependencyIndexes: file_demo_model_v1_tasks_proto_depIdxs,
		MessageInfos:      file_demo_model_v1_tasks_proto_msgTypes,
	}.Build()
	File_demo_model_v1_tasks_proto = out.File
	file_demo_model_v1_tasks_proto_goTypes = nil
	file_demo_model_v1_tasks_proto_depIdxs = nil
}