# openapi/user.json
{
  "components": {
    "schemas": {
      "demo.model.v1.Task": {
        "description": "Task 是任务数据；时间使用 RFC 3339 JSON 字符串。",
        "properties": {
          "completed": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "demo.model.v1.TaskEvent": {
        "description": "单帧 data 的 JSON 结构，对应 api/demo/model/v1/tasks.proto 的 TaskEvent；不是整个 HTTP 响应的结构。",
        "properties": {
          "eventId": {
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "type": "string"
          },
          "task": {
            "$ref": "#/components/schemas/demo.model.v1.Task"
          },
          "type": {
            "enum": [
              "task.created",
              "task.updated",
              "task.deleted"
            ],
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "type",
          "task",
          "occurredAt"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Go + Buf + gRPC-Gateway 示例。任务接口由 proto 生成；SSE 与健康检查通过受控文档片段合并。数据只保存在内存，重启后清空。",
    "title": "Proto API Lab · 用户端",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/user/v1/events": {
      "get": {
        "description": "真实的 text/event-stream 长连接。先发送 ready，之后发送 task.created、task.updated、task.deleted；每 10 秒发送注释心跳。支持 Last-Event-ID 或 last_event_id 查询参数重放最近 128 条事件。事件历史过期或服务重启后游标超出范围时发送 reset，客户端重新拉取列表。Swagger 的 Try it out 不适合持续流，请使用首页的实时面板、EventSource 或 curl -N。",
        "operationId": "SubscribeTaskEvents",
        "parameters": [
          {
            "description": "最后收到的非负整数事件 ID；优先于查询参数。",
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            }
          },
          {
            "description": "方便浏览器首次连接时指定游标。",
            "in": "query",
            "name": "last_event_id",
            "required": false,
            "schema": {
              "pattern": "^[0-9]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "example": "retry: 1000\nevent: ready\ndata: {\"latestEventId\":\"0\",\"replayed\":0}\n\nid: 1\nevent: task.created\ndata: {\"eventId\":\"1\",\"type\":\"task.created\",\"task\":{\"id\":\"task-000001\",\"title\":\"体验 SSE\",\"completed\":false},\"occurredAt\":\"2026-09-14T00:00:00Z\"}\n\n",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SSE 流，响应持续保持打开"
          },
          "400": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "游标不是合法的非负整数"
          }
        },
        "summary": "订阅任务变化（SSE）",
        "tags": [
          "实时事件"
        ],
        "x-sse-events": {
          "ready": {
            "description": "订阅已建立；携带 latestEventId 和 replayed。"
          },
          "reset": {
            "description": "游标过期；重新拉取当前端的任务列表以同步状态。"
          },
          "task.created": {
            "$ref": "#/components/schemas/demo.model.v1.TaskEvent"
          },
          "task.deleted": {
            "$ref": "#/components/schemas/demo.model.v1.TaskEvent"
          },
          "task.updated": {
            "$ref": "#/components/schemas/demo.model.v1.TaskEvent"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}